refactor(backend): Removed unused code
and moved a bit around
