--- src/main/java/ecostruxure/rate/calculator/bll/MissingCurrenciesException.java
@@ -1,7 +1,0 @@
-package ecostruxure.rate.calculator.bll;
-
-public class MissingCurrenciesException extends Exception {
-    public MissingCurrenciesException(String message) {
-        super(message);
-    }
-}
