--- src/main/java/ecostruxure/rate/calculator/bll/service/CountryService.java
@@ -1,23 +1,0 @@
-//package ecostruxure.rate.calculator.bll.service;
-//
-//import ecostruxure.rate.calculator.be.Country;
-//import ecostruxure.rate.calculator.be.Geography;
-//
-//import java.io.IOException;
-//import java.util.List;
-//
-//public class CountryService {
-//    private final ICountryDAO countryDAO;
-//
-//    public CountryService() throws IOException {
-//        countryDAO = new CountryDAO();
-//    }
-//
-//    public List<Country> all() throws Exception {
-//        return countryDAO.all();
-//    }
-//
-//    public List<Country> allByGeography(Geography geography) throws Exception {
-//        return countryDAO.allByGeography(geography);
-//    }
-//}
