+++ src/main/java/ecostruxure/rate/calculator/controllers/CurrencyController.java
@@ -1,11 +1,9 @@
 package ecostruxure.rate.calculator.controllers;
 
 import ecostruxure.rate.calculator.be.Currency;
-import ecostruxure.rate.calculator.bll.service.CurrencyService;
+import ecostruxure.rate.calculator.bll.currency.CurrencyService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 
 @RestController
