+++ src/main/java/ecostruxure/rate/calculator/controllers/GeographyController.java
@@ -1,13 +1,11 @@
 package ecostruxure.rate.calculator.controllers;
 
 import ecostruxure.rate.calculator.be.Geography;
-import ecostruxure.rate.calculator.bll.service.GeographyService;
+import ecostruxure.rate.calculator.bll.geography.GeographyService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/api/geography")
